requirements.txt
streamlit
pandas
matplotlib
sklearn.linear_model
gspread
oauth2client.service_account
plotly
os 
